create lexer
Lexer performs lexical analysis and works with parser to compile a code. It recognizes 4 types of token: words, integers, real numbers, real with an e., special characters, and column equal. DFA needs to be imported and read into a 2D array (TwoDArray[] in my case) for the lexer to work.